+++ src/main/java/io/github/odeivissonsantos/repositorys/BarbeiroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.odeivissonsantos.models.BarbeiroModel;
 
-public interface BarbeiroRepository extends JpaRepository<BarbeiroModel, Integer> {
+public interface BarbeiroRepository extends JpaRepository<BarbeiroModel, Long> {
 
 }
