+++ src/main/java/io/github/odeivissonsantos/repositorys/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.odeivissonsantos.models.ClienteModel;
 
-public interface ClienteRepository extends JpaRepository<ClienteModel, Integer> {
+public interface ClienteRepository extends JpaRepository<ClienteModel, Long> {
 
 }
