+++ src/main/java/io/github/odeivissonsantos/repositorys/ServicoPrestadoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.odeivissonsantos.models.ServicoPrestadoModel;
 
-public interface ServicoPrestadoRepository extends JpaRepository<ServicoPrestadoModel, Integer> {
+public interface ServicoPrestadoRepository extends JpaRepository<ServicoPrestadoModel, Long> {
 
 }
